+++ app/Filament/Resources/Members/Pages/EditMember.php
@@ -16,4 +16,9 @@ class EditMember extends EditRecord
             DeleteAction::make(),
         ];
     }
+
+    protected function getRedirectUrl(): string
+    {
+        return $this->getResource()::getUrl('index');
+    }
 }